[core] Improve 'ContinueAnyExistingRequest'
Now we only try to continue an existing request if the
peer has been unable to select one normally.

Additionally, we only continue other peer's pieces if
the peer in question has zero pending requests. This
is slightly inefficient in terms of pipelining, but
it's probably better that a peer finish it's queue of
up to 300-500 pieces *before* it starts stealing pieces
from other peers.