[core] Trigger EndGameMode when 'ContinueAnyExisting' gets no pieces
The new semantics make it better to enter endgame mode when
ContinueAnyExisting returns null rather than when PickPiece
returns null.

We intend for PickPiece to return null when all pieces have
been assigned, and at that point we are OK with getting peers
to share pieces.

This should be a settable behaviour as we'd rather take the
performance hit when streaming media rather than be stuck
trying to get a whole piece from a slow peer.