[core] Be a little more threadsafe with ITorrentFileStream
Change around ownership of the Semaphore used to control
access to the streams (again).

There was a subtle race condition before where it was theoretically
possible for a thread to be waiting to write a piece of data, and
another one waiting to close the stream, and as a result
the first thread would begin execution using a Stream which
has already been disposed.

Now the first phase in any interaction with the stream is to
enter the Semaphore, which then allows the stream to be used,
or created, in a threadsafe way.