[core] Lock fewer times when iterating the main loop
Use two queues to allow tasks to be processed with
fewer locks.

By swapping the queue object we can dequeue every
message in the old queue object without needing
to take further locks. When the old queue is empty,
we can swap it again and start processing tasks
in the new queue.