[core] Use DiskManager to safely read the torrent data
The DiskManager implementation may, or may not, use a
memory cache. As per comment in the code - using a
FileStream and assuming the data we need is on-disk
is unsafe and will (probably) cause problems.

The recent switch to use a memory cache by default has
made this happen!

Swap this code to use the internal DiskManager to read/write
data safely.

Fixes https://github.com/alanmcgovern/monotorrent/issues/359