[core] Update MemoryWriter to be async safe
We cannot iterate over 'CachedBlocks' and use
an await unless we write our code so that it
is aware that other tasks may execute while it's
flushing.

When this was originally written the DiskWriter
was single-threaded and synchronous. This meant
nothing could modify the 'CachedBlocks' list
while the writer was iterating and flushing.

By moving to an async workflow, it is possible
for other tasks to execute and modify this list.

Partial fix for https://github.com/alanmcgovern/monotorrent/issues/360