[core] Flush the DiskManager before reading data
It would be incredibly complex to support a scenario where some
parts of a single read are in the in memory cache and some are
on disk. We might have to read 1 block from memory, the next from
disk, then the third could be in memory.

This may be a tad inefficient, but it makes it trivial to ensure
we get correct data.