[core] Improve flushing for the tests
The tests don't necessarily run a main loop which ticks
the DiskManager to process queued writes.

This change should be tiny perf improvement in realworld
scenarios as we only queue the flush task if there are
are queued writes which need to be flushed.

Additonally, we attempt to process enqueued tasks
immediately to see if we can complete the flush
right away.