[core] Fix some racey tests
Don't bother creating a real Udp based DhtListener for the tests
as we never acutally use it.

The tests which use Dht use a ManualDhtEngine, or we test
the specific aspects of Dht which are relevant. 99.9% of
tests do not need to bind an actual Socket.

The tests which were failing were failing asynchronously
in a way which didn't really matter. The engine catches
and handles errors stemming from announcing to trackers
already, and so the exception was masked.