[core] High priority pieces can be downloaded from multiple peers
This is less efficient as data can arrive out-of-order which
means we'll write data to disk, then read it back from disk
in order to compute the piece's hash.

However for a streaming torrent it's more important to get
data *quickly* when it's in the high priority set.

The memory cache should reduce the odds of actually flushing
the data to the hard disk though.