[core] Disconnect peers who time out fulfilling piece requests
If we request a piece and it is sent at less than 1kB/sec we
should simply disconnect the peer. They're too slow to be of
any practical use.

This simplifies the code as there's no such thing as a 'timed out'
piece request anymore. Either peers are actively fulfilling requests
or they are disconnected for failing to do so. Partial pieces from
disconnected peers are automatically treated as abandoned.