[core] Mark pieces as abandoned, not blocks
It's simpler all round to mark pieces as abandoned instead of
monitoring this on a block-by-block basis.

Previously a block could be 'abandoned' if the request associated
with that block timed out. Timed out requests don't exist anymore
so the logic to handle this doesn't need to exist.

If a peer 'times out' it is disconnected. At that point every
piece it contributed to can be marked as 'abandoned' and so
any peer can request blocks from it.