[core] Eliminate a harmless NullReferenceException
It's not impossible for the connection to be disposed
before we reach the Socket.Connect call.

Update this so we do not trigger a NullReferenceException
when this occurs to avoid this looking like a bug,

ObjectDisposedExceptions are converted into 0 when
send/receive is invoked which continues the same
behaviour as before.