Refactor the IPiecePicker interface for easier implementation/usage
The idea is to:
1) Make it easier for the engine to indicate when a peer should be
used to download a piece without other peers sharing, or whether
the piece should be downloaded with other peers sharing.

2) Remove the need to think about the fast peer extensions. Cancellation
will be handled by the engine itself.

3) Remove a bunch of stuff from the IPieceRequester API

4) The 'Piece' class was public API but did not have a public
constructor. This made it harder to use. It's now internal
API and the two pieces of information we needed are passed out
separately. All we need to know are if the piece has been
fully downloaded, and if it has we want to know which peers
sent us the blocks.

5) IPiecePicker is a little more flexible and clearer for external
implementators. When a peer disconnects we call 'AbortRequests' which is
clearly documented as meaning 'cancel all unreceived pieces and none
will be requested again'.

'CancelRequests' takes a start/end index so a streaming picker can
determine whether or not to cancel pieces. When a request is cancelled
the list of cancelled pieces is returned so CancelMessages can be sent.
This part probably needs alteration as there are *two* ways to cancel:
The internally implemented StreamingPicker and/or LocalStream can do
it because they have access to everything. 3rd party implementations
may also need to do this!

The 'RequestRejected' message is now propagated into the PiecePicker
as distinct API so no-one has to figure out if a 'CancelRequest' message
needs to be sent. Requests cancelled using the Cancel API would need
CancelMessages, but requests cancelled by 'RequestRejected' do not
need a CancelMessage as the remote peer has already acknowledged and
rejected the piece.