[core] Only cancel pieces if we seek to data we don't have
Some libraries/apps like to seek backwards in the stream
even though they've already read the data. In this scenario
we do not want to cancel any outstanding requests as we
already have the data in the high priority set.

Enforce this by modifying the code so we only cancel requests
if any of the data in the new area is *not* already downloaded.