More picking tweaks (#394)
Next round of refactorings. I'm a bit happier with this approach.

IPiecePicker - implement this to control exactly which block is requested from the available pieces. Custom implementations would, ideally, not be necessary as StandardPicker should be robust enough to cover all use cases.

IRequestManager - implement this to control how methods on IPiecePicker are invoked. Example usecases would be:

    Implement an IRequestManager which allows duplicate requests to be made for high priority pieces.
    A customised way of entering 'endgame mode'
    Customise exactly which peers get assigned blocks from each piece. You could enqueue important pieces from peers which have a high transfer rate.

Some of this will be tweaked a bit further :)