Replace StreamingPiecePicker with StreamingPieceRequester
Move the specialised streaming piece picker logic into the new StreamingPieceRequester.

This simplifies everything overall. The logic to determine which piece/block should be requested is now completely separate from the logic to actually pick a piece/block.

With this change the prioritisation of pieces is controlled within StreamingPieceRequester. It is in charge of issuing duplicate requests for pieces in the ultra-high priority set, requesting the high-priority set sequentially, setting limits on the maximum number of pending requests enqueued with peers, and also enqueuing important requests with peers which have previously offered high transfer rates.