[core] Add piece requests on a regular cadence
If there's a peer we *should* be requesting pieces from but
are currently not requesting any, try to request some.

During regular downloading the engine will call
AddRequests (IPeerWithMessaging, IReadOnlyList<IPeerWithMessaging>, BitField)
as soon as the peer sends an Unchoke message, and also as soon as any
PieceRequest is fulfilled. As such, we should never have a scenario
where a peer which is unchoking us will not have pieces requested.

This can happen during EndGame mode as we will have at most 2 or 3
outstanding requests per block. In this scenario we may end up
failing to request pieces from peers which are unchoking us, so this
will ensure those peers are kept in the rotation and we will,
eventually, request blocks from them if needed.