[core] Fix a bug when the MemoryWriter capacity is tiny
If the MemoryWriter capacity is smaller than the size of
the block being written, it should short-circuit and write
the data directly to the underlying PieceWriter.

In this kind of scenario it's quite likely there'll be no
data stored in the MemoryWriter (so we'd IndexOutOfRangeException
if we tried to flush a block of data), but even if we did so
the new block of data wouldn't fit in the MemoryWriter anyway.