[core] Use the base version to determine the announced version
This isn't ideal, but it's good enough for now. It's pretty
much mandatory to update the version.txt file when issuing
new releases to ensure they're consecutive. As such we should
just use that number when setting the version we announce
on the bittorrent protocol, and the auto-incrementing verison
number will be used to disambiguate preview releases etc.