[core] Make IncrementalHashing threadsafer
If two blocks from the same piece are received at the same time
and are processed in sequential ticks of the main loop,
it's possible they'll both be executing Mode.WritePieceAsync
at the same time.

In turn, this means they'll both invoke 'DiskManager.WriteAsync'
and queue the block of data to be written to disk. If we received
the second last and last blocks in a piece, the DiskManager.WriteAsync
calls can complete in the opposite order, which in turn can mean
the threadpool thread which is attempting to compute the
incremental hash for the second last piece may race with the call
to 'GetHashAsync', and we could miscompute the incremental hash.

The window for this is incredibly small, but the use of the
exclusive semaphore should eliminate it completely.

This would not cause corrupt data to be written to disk, it would
simply make the engine believe it had downloaded data which failed
it's hashcheck.