[core] Flush the disk cache if the capacity changes
This should rarely, if ever, change during runtime. In the
event it does change it's probably best that we flush all
pending data. This makes some lifecycle management a
little easier.

For example if the old size was 5MB and the new size is 0,
we do not want any pieces to remain in the cache. Flushing
everything saves us the hassle of determining exactly
how many pieces should be left.