[core] Set disk cache size to 0 by default
I'm pretty confident that we can safely keep MemoryWriter available
for people who wish to slightly improve performance and try out
this experimental feature.

Default it to 0 in size so it's effectively off. Someone can set
this to a reasonable value (5MB?) if they wish to help test it
in the real world.