[core] Introduce a proper IBlockCache API
Rather than integrating a memory cache as an IPieceWriter, we can
instead integrate it as it's own distinct thing. This can greatly
simplify the implementation.

The engine downloads pieces in blocks of length 16kB. These blocks
are hashed incrementally, and somewhat sequentially, until the
hash for the entire piece is ready.

If we specialize the cache so that it operates on 16kB blocks then
it's far easier to implement and manage. This is represented by
IBlockCache.

IPieceWriter simplifies the task of reading/writing data to
individual files in a torrent. We maintain this behaviour so
implementers of IPieceWriter do not have to worry about converting
between 16kB blocks and ITorrentFiles/offsets.