[core] Improve incremental hashing with the memory cache
Incremental hashing now fully guarantees the invariant
that a block which is incrementally hashed will never be
stored in the in-memory cache.

It also guarantees that a block will never be incrementally
hashed after it has been stored in the in-memory cache.
Data placed in the in-memory cache will always be read
when the 'GetHashAsync' method is invoked when the final
block has been recieved, and this in turn will ensure
all blocks related to the piece are written to the
IPieceWriter before the piece is marked as 'downloaded'.

Note: The old behaviour of counting the number of blocks
which have been written has been restored too. If block
N and N+1 arrive sequentially, there's no guarantee which
order the async methods will return when MemoryCache.WriteAsync
is invoked.

As such, we cannot rely on the 'pieceComplete' boolean to
know the final piece has been written. It's only sufficent
to inform us the final piece has been received.