[core] Ensure that the non-null peersInvolved is used
This is cached because we know the version passed into
the method is not guaranteed to be non-null.

However, we do have a guarantee that it will be non-null
when the final block for a given piece is received, so
by the time all blocks are written to disk we will have
a non null entry in the cache.