[core] Make UpdateSettings a normal 'async' method
It's probably best, long term, to be consistent. Make this
method async and ensure it proxies to the IO loop. While
the code in here is (currently) threadsafe, that may not
always be the case.

Also clean up some of the tests. They're a little complex
now after the recent factorings so hopefully this change
will make their intent clearer.

When queueing up reads in DiskWriter *any* of them could
be the first to complete. Some will be sequential, some
will be in parallel.