[core] Address some potential races in the MemoryCache
There used to be the potential for a data race. The issue
is that there'd be a gap where a particular BlockInfo was
removed from the MemoryCache but was not yet written to
disk. In this situation a 'IBlockCache.ReadAsync' invocation
will not find the block in memory, but if it attempted to
read it from disk it may not yet have been written.

This change ensures that the data has been written to disk
before it's removed from the memory cache.