[core] Rename some settings for clarity
If it's supposed to be a directory, append 'Directory' to
the property name.

This replaces usages of 'Path' and 'Folder' which are
somewhat ambiguous. Is a path a file or a directory?

The old properties are being maintained, temporarily,
but marked as Obsolete. This will allow a slightly easier
migration for people trying to betas.