[core] Change how Streaming is used
Move this to ClientEngine.AddStreamingAsync. When this overload
is invoked we'll now populate the TorrentManager.StreamProvider
property, and from there you can create streams for any of the
files in the torrent.

It simplifies things a little as regular downloading and streaming
are how handled in nearly exactly the same way - you call a
method on ClientEngine to create the TorrentManager.

In addition to this, it's now easier to guarantee that the
PiecePicker isn't changed when Streaming mode is activated.
The built in 'StreamProvider' creates 'LocalStream' objects
which have a dependency on the actual PiecePicker being used
so they can cause specific pieces to be prioritised.