[core] Make some state non-static
If you create two DhtEngines then message ids should not be
shared. This wouldn't have been a problem as IDs are guaranteed
to be unique (as much as is possible when you only have 2 bytes).

However, it's probably better to make this non-static and reset
things when the Dht Engine is stopped/started.