[core] Use a proper way to change file priority (#416)
Don't allow people to set this directly via public API.
It's much more efficient for the library to handle updates
via an async method.

Firstly, it removes the race condition where the user could
be trying to update the priority while monotorrent is accesing
it. Accesses to this property have always had to be carefully
made as the value could change immediately after it was
read. Now these concerns are eliminated.

Secondly, it provides a convenient place to update some
state in the library. For example calculating partial progress,
and also when updating the priority based piece picker.