[core] Make the 'BitField' mutating methods internal (#422)
The 'BitField' class is supposed to be readonly in every scenario
where it's exposed to users.

After a fair bit of thinking, the best way to enforce this is to
make the mutating methods internal. This will allow the monotorrent
to update the bitfields as data is modified, while ensuring users
cannot accidentally corrupt data by using the mutating methods
unexpectedly.

If, for some reason, users of monotorrent need a mutable bitfield
then that can be exposed. However a better choice may be for them
to use the built in System.Collections.BitArray class.

Addresses https://github.com/alanmcgovern/monotorrent/issues/418