[core] Add a mutable bitfield to the public API
This now gives us a 'guaranteed' immutable BitField,
and also a Mutable version. Most things internally
instantiate a MutableBitfield but expose it to the
user as a plain old 'BitField'. This reduces the risk
of users accidentally doing the wrong thing and mutating
one of the bitfields the engine is managing.