[core] Add the ability to delete data when removing
If a magnet link is used to create a TorrentManager then
the metadata is written to the cached directory after it
is successfully downloaded.

If a filepath is used, then the file will be copied to
the cache directory.

If a raw 'Torrent' is used, the metadata stored by the
torrent is written to the cache directory too.

The path to the metadata is exposed by the TorrentManager.

'ClientEngine.RemoveAsync' now has an overload to control
which data should be deleted when the torrent is removed.
You can remove nothing, the cache data, the downloaded data,
or all of the above.