[core] Error correctly if the same torrent is loaded twice
DownloadMetadataAsync was a secret backdoor to accidentally
loading the same torrent twice, but only if you used the
overload which took a 'MagnetLink'.

Fix the issue by using the InfoHash to ensure torrents
weren't loaded twice, as the 'Torrent' property is null
for magnet link downloads.

Adds a test covering this.
Fixes https://github.com/alanmcgovern/monotorrent/issues/452