[core] Use an immutable approach to overriding core functionality
Require the user of the library to create the 'Factories' object
before initializing the engine, and ensure 'Factories' is immutable.

If someone wants to enable/disable something core, like proxy support,
then it's far easier to do that safely/reliably if they also instantiate
a new ClientEngine object.

These are things which should not change for the lifetime of a torrent,
unlike the more mutable EngineSettings or ClientSettings.