[core] Move some tracker stuff around again
Also split ITrackerConnection out of ITracker.

Internally a tracker is just a tracker, the only thing that actually
changes is the mechanism behind sending an 'Announce' or 'Scrape'
request, so move this into an ITrackerConnection interface and
simplify things internally. The 'connection' used to contact a
tracker is independent of how the engine needs to handle a
successful, or failing tracker.