[core] Add standardised caching/resetting to PeerMessage
Previously PieceMessage had specialised support which allowed
the underlying buffer to be freed at the correct time. Change
the code around a bit so all PeerMessages can benefit from a
deterministic lifespan.

This allows PieceMessage buffers to be freed at the appropriate
time and also allows for caching/reuse of PeerMessages.

Let's give this a shot!