[piecepicking] Some small GC improvements
If we retain/reuse Piece objects we eliminate that last of the
per-piece allocations.

tbh there's precious little, if anything, left to really optimise
in terms of memory allocations. Downloading a 3GB torrent can
allocate as little as 50MB in total (all retained objects, and all
collected garbage).

The next big thing is to complete bep52 support!