[bep52] Support validating torrents using V2 hashes
Next phase of downloading bep52 compliant torrents - complete.

Now that the data is parsed, validated and available, we just
need to change the validation code to use SHA256 hashes if/when
a V2 connection is active.

Note: There's no guarantee the piece layers will be stored with
the torrent metadata, in which case we'll have to download that
data dynamically from peers.

As such, we will need a mutable datastructure so we can update
the piece hashes on-the-fly as new data arrives and is validated.
Data loaded from the .torrent is intended to be *immutable*.