[core] Correctly detect V2 torrents when populating infohashes
The SHA256 and SHA1 infohashes are always computed as it's not
easy to predetermine whether a torrent will be v1, v2 or hybrid.

We'd have to double-parse the torrent to do that reliably, which
would require the 'Stream' input to be buffered/copied, or the
Stream would have to support seeking.

As such, only populate the SHA256 hash (in the end) if the torrent
files have a PiecesRoot property which is non-empty. This is
required for v2 torrents.