[client] Fully wrap IDhtEngine
ClientEngine automatically starts/stops many things, including the
Dht engine registered with the engine. It also starts/stops the
associated listener whenever the port changes, and registers it
with upnp/nat-pmp.

Some people were confused in the past and tried to start/stop the
Dht engine manually. Let's wrap those APIs in a way which eliminates
that confusion.

Announcing is also handled internally so no-one should need to mess
with those methods either.