[core] Load hybrid torrents correctly
Hybrid torrents aren't fully supported (yet). To support these the
engine will load them as either V1 or V2, and that is accomplished
by ensuring only the V1 hash, or the V2 hash is set after loading
the torrent.

In addition to this, we should also ensure that the PieceHashes
object is not set to non-null for both hashing schemes. A v2 only
torrent should only have a V2 hashes object.