[bep52] Modify the API to support dynamically retrieving Bep52 hashes
The torrent metadata is no longer the source of truth for piece
hashes. Piece hashes are optionally part of a bittorrent v2 (bep52)
compliant torrent. In the event they're missing, they can be requested
from peers after the metadata has been retrieved.

One invariant I wish to maintain is that the 'Torrent' object should be
completely immutable.

As such, remove the 'IPieceHashes PieceHashes' property and
expose a public method instead. This method has an overload where you
can supply a Dictionary of piece hashes, keyed by
'ITorrentFile.PiecesRoot'. This way an IPieceHashes object can be
created after hashes are fetched from another peer and those hashes
have been validated as being correct.

'TorrentManager' is mutable and stores the latest version of the
PieceHashes object, allowing it to be updated when the data has been
fetched.