[core] Better read-only bitfield support
By having 'MutableBitfield' be a subclass of 'BitField' it's possible
for users of the library to upcast to 'MutableBitField' and do the wrong
thing.

Follow the pattern of Memory<T>/ReadOnlyMemory<T> and create BitField
and ReadOnlyBitField.

BitField can be converted to 'ReadOnlyBitField' via an implicit cast,
and you can create a BitField from a ReadOnlyBitField by using the
appropriate constructor.

This should make things a little bit clearer.