Pick anything, everywhere, all of the time!
Move piece picker so it relies on two things.
1) The concept of a 'Piece'. Each piece is a thing which can be sharded
   into blocks.
2) Blocks. Each piece consists 1 or more blocks.

The idea is that a 'piece' is an independently verifiable chunk of data
whereas a 'block' is not (necessarily) independently verifable.

The four examples we have today are:
1) Fetching the torrent metadata. This is fetched in non independently
   verifable chunks.
2) Fetching hash metadata. This is fetched in independently verifiable
   chunks
3) A 'Piece' in BitTorrent V1 is 32kB or larger, and is fetched in
   chunks of 16kB. These chunks are non independently verifable.
4) A 'Piece' in BitTorrent v2 is 32kB or larger, and is fetched in
   chunks of 16kB. These chunks can be independently verified, but
   typically will be verified at the Piece level.

Requesting all of these requires the same kind of logic around
requesting from multiple peers, re-requesting pieces which don't arrive,
timing out requests, validating the responses etc etc.