[piecewriter] Add a cache policy so reads can be cached
This will be useful when creating torrents. In this scenario
we want to be able to read a piece into memory, hash it using
both SHA1 and SHA256, and then return the bittorrent v1 and
v2 hashes for that piece.

Afterwards, we need to compute the SHA1 and MD5 hashes for the
individual file. This would normally require re-reading the
exact same data from disk, but now we can fetch it from memory.