Update TorrentCreator to create V2 torrents
Delete all the bespoke hashing logic and replace it with
DiskManager.GetHashAsync.

DiskManager is already smart enough to calculate v1 and v2
hashes for a torrent, so why not just use that to generate
the hashes when *Creating* a torrent.

If either MD5 or SHA1 hashes are requested, leverage the
'read' cache to allow passing the data through an
MD5/SHA1 hasher after the primary V1 and V2 hashes are
generated.