[core] Hide 'Peer' internally
Rely on PeerInfo as the public interface for peers which have
no active connection.

PeerId is still the public interface for peers which do have
an active connection.

Rename the 'IPeer' interface in the piece picker to be a little
more obviously related to requesting pieces.

This allowed one of the (many) 'peer' named classes to be made
internal.