[core] Refactor IPiecePicker to return bools and use out variables
This makes things a bit more consistent. The most common usecase
of this was to check if the nullable return value was non-null and
then do something with it. This can now be collapsed into a 1 liner
as the return value is just a plain bool, with the PieceSegment being
returned as an out variable