[piecepicking] Use a HashSet rather than out IList in the API
If we want a collection of the distinct peers which a piece was
fetched from, we should just use a HashSet. It's what it does.

Additionally, rather than requiring the PiecePickers to instantiate
their own hashset, just pass one in. This opens the door to an
optimisation where the HashSets can be cached/reused.