[samples] Just load the filepath directly into the engine
It's better for samples to do things in the blessed way so that
people who replicate samples get the best experience.

With this method, the 'real' .torrent file is copied to the cache
instead of forcing the engine to synthesize a fake torrent file
using random fields from the Torrent object itself.