[core] Properly clean up disconnected peers from the UnchokeManager
When peers are disposed, their message queue is also disposed, which
forbids new messages from being enqueued.

It's possible that some disconnected peers remained in the unchoke
manager which could lead a Choke, or Unchoke, message being sent
to a peer whose connection has been closed.

Added a test too.

This is awkward to test as you need to have more choked, but interested,
peers than you have slots for, and for a peer to disconnect, and for
this to happen in between unchoke reviews.