[bep52] Set the v2 infohash when the first file is empty
The previous logic did not take into account the fact that
the first file could be empty, in which case it will not have
a piecesroot.

There is a better way to identify v2 metadata without relying
on the existence of 'pieces root' for arbitrary files, so use
that method instead.