[core] Sort files before creating hybrid or v2 torrents
As the underlying storage is a bencoded dictionary, whose
keys must be sorted, you *must* sort the list of files which
will be placed in a hybrid or v2 torrent prior to hashing
them, otherwise the order of the files will be flipped and
piece hashes won't match anymore.

This doesn't appear to be mentioned in the spec, but is
called out in a code comment in the sample torrent creator
linked as part of the bep52 spec.

            # the directory entries must be processed in lexicographic order so that
            # the files list matches the bencoded ordering of the file tree

I went hunting as soon as i realised i didn't know how to
make this test pass as the files were not in a bencoded list.

Fixes https://github.com/alanmcgovern/monotorrent/issues/563

Test case written by @borigas