[core] Remove a sanity check which doesn't make sense anymore
Both BEP47 (padding files) and BEP52 (BitTorrent V2) introduce the
ability to request blocks which are less than 16kB in the 'middle'
of the torrent.

Also, BEP52 requires all compatible clients to close the connection
if any peer requests > 16KiB of data. As this is pretty standard,
and the existence of clients which request larger chunks than that
is exceedingly rare, just implement that.