Files with no content should have no piece root.
WHen checking the hash of piece 0, we should skip past files
with a length of 0. This is important as if the file is smaller
than PieceLength in size, the 'PieceRoot' of the file is used
instead of the piece hashes, and so we need to ensure we pick
the pieces root from the file which has the actual content.