[core] Remove PieceRequesterSettings.IgnoreChokeState
This was a hack which shouldn't be in the public API.
There are (now!) two cases where we need an IPiecePicker
which ignores whether or not a peer reports as 'choking'.
This is:
    1) When fetching metadata
    2) When fetching piece hashes for bittorrent v2

In both of these cases the value of 'IRequester.IsChoking'
is irrelevant as peers who are choking will still
fulfill a request for metadata, or for piece hashes.

The 'correct' way of supporting this is to just implement
an IRequester wrapper which harcodes 'IRequester.IsChoking'
to false.

After doing that, and after eliminating the hacky setting,
a further change had to be made which is to ensure a bitfield
of the correct length, with all bits set to true, is
passed to the 'AddRequests' method so that it's clear that
any of the blocks can be requested from these peers.