[core] Ensure timeouts apply for MetadataMode and PieceHashesMode
If a block is requested from a peer when fetching metadata, or
when fetching the piece hashes of a bittorrent v2 torrent, we
need to ensure the '45 second' timeout is applied to these
requests, just like they are applied for regular piece
requests.

Additionally, whenever a request is enqueued with a peer
we should reset the 'LastBlockReceived' timer so that we
do not unfairly disconnect a peer for not fulfilling a
request within the timeout period simply because we
have *literally* just requested the block.