[core] Fix an overread when streams have more data than asked for
The ZeroStream implementation always fills the passed-in buffer
with zeros, as any well behaving Stream implementation should.

The intention was for *at most* 'DataStreamCount' bytes to be read
from the stream, so just slice the buffer like is done everywhere
else in the codebase so it's the correct size.

Sorted!