[core] Add ipv6 support for udp trackers
Each tracker can npw have one, or more, connections. When a tracker is
announced to, each connection will be used to connect the tracker.

Typically this should be used to do an ipv4 *and* ipv6 announce/scrape
to a tracker (where supported).

A tracker is treated as online if any of the connections succeeds.