[core] Use a per-tracker random key for announcing
Change the key generation logic a little and share it between
all tracker implementations.

In short - the key is a random (positive) 32bit number. It's unique
per-torrent and per-tracker, so someone cannot correlate the same
client across different trackers or torrents.

However the same key is sent for all connections for a given
tracker/torrent pair. i.e. an ipv4 and ipv6 announce to the
same tracker, for the same torrent, will send the same key.
If a hybrid torrent is being downloaded, then the key is sent
for both the V1 infohash and V2 infohash.