[core] Use SocketsHttpHandler everywhere
For older versions of the .NET framework, use the
'StandardSocketsHttpHandler' nuget as a poly-fill to provide the
capability to bind to either an IPv4 or IPv6 local endpoint in
order to correctly implement the multi-announce strategy to support
IPv6 trackers.

For trackers which are routable across IPv4 and IPv6, clients are
supposed to announce using both IPv4 and IPv6.