[core] Refresh the 'AmInterested' status regularly
There are several cases where the 'AmInterested' status can change
for a given peer, and it's not easy to provide an event for each
of these.

For example, the 'interested' status is dependent on each bitfield
added as an 'IgnoringBitfield' to the piecepicker. These bitfields
can change arbitrarily. This async check will ensure that things
will refresh regularly.

As updating the priority of a file does have a pretty clear user
action associated with it, special case this to cause an immediate
refresh of the 'AmInterested' status.