[core] Simplify how excess streams are closed
I had many concerns about the prior approach to closing excess
streams due to how it used a while loop and also async/await.

To keep things simple, make MaybeRemoveOldestStreams a non-async
method, so it's a bit easier to reason about what's happening.
The method now chooses the stream which should be disposed,
cleans up the internal state, and then queues that stream for
async disposal.

The engine does not have a strict limit on total open handles,
and it makes no promises about *precisely* how many filehandles
can be open at a given point in time, so the fact the dispose
is slightly async is still fine. If someone has an actual
hard limit on the exact number of file streams which can be
opened, then this delay is just one of many issues they'll
encounter.