[core] Fix several bugs with 'EngineSettings.UsePartialFiles'
This adds several additional tests covering aspects of how this
feature was intended to operate.

The 'Complete' and 'Incomplete' paths are now generated using
a uniform method. The choice of whether or not to use the
'Complete' or 'Incomplete' path as the current path is
dependent on whether or not 'UsePartialFiles' is on, and
if the file is complete.

This should keep everything in sync when toggling UsePartialFiles
on and off, and also if the user uses 'MoveFileAsync', or
'MoveFilesASync' to move files.