[core] Ratelimiters should allow requests larger than chunk size
If the ratelimiter is set to download at 15kB/sec or less, the
amount of data in the ratelimit will always be less than the size
of one block. In this case, we should allow the request to succeed.

Once there's negative data left in the allowance, deny requests.
As the ratelimiter is topped up over time, future requests will
be unblocked.