[core] Tweak ratelimiting a little better for low rates (#618)
* [core] Tweak ratelimiting a little better for low rates

Fix another issue with a ratelimit of 1 byte/second. Hopefully
no-one was actually trying to use a limit this low.

Additionally, simplify this logic now. Rate limits are much
easier than they used to be as the engine works in bytes, not
'blocks' or 'chunks'.

Back in the day there had to be estimates made for the amount
of data which would be received for a given 'ReceiveAsync' call.
Now we're just counting the raw number of bytes sent/received,
or read/written, so we don't need complicated math to account
for over/under-shoot.