[core] Add improved logging for unexpected peer_ids
If a peer sends a HandshakeMessage with a 'peer_id' that *does not*
match the 'peer_id' it sent to the tracker - log both the expected
and actual peer_ids.

Helps diagnose https://github.com/alanmcgovern/monotorrent/pull/614

It's probably a normal bittorrent client which is just randomising
the peer id on every connection attempt.