[core] Try harder to send a valid port to the tracker
The main part of this change is that the actual IPEndPoint
which the listnener has been bound to is not the sole source
of the 'port' value which should be sent to the tracker.

If the listener has been configured to bind to a specific
port, but has (at the point in time when the announce is
sent) not successfully bound to that port, the preferred
port should be sent to the tracker.

If the user has configured the listener to port '0' *and*
the listener isn't running, then we have a problem as 'port'
is a required parameter and must be sent. In this scenario,
send an invalid port and the tracker should still return a
list of peers to connect to.